Return to using utils without core/ folder
